[dsbx] extract shared rewrite policy
Move MITM request policy checks and DSEC placeholder substitution out of the byte-level HTTP/1.1 rewriter into a shared rewrite_policy module. The h1 path still calls the same policy with Host-header authority, preserving the existing framing and deny behavior.
